// mypivot.sources/mypivot4-fe/projects/mypivot4-fe/src/app/model/capitolo-rt.ts
/*
 *     MyPivot - Accounting reconciliation system of Regione Veneto.
 *     Copyright (C) 2022  Regione Veneto
 *
 *     This program is free software: you can redistribute it and/or modify
 *     it under the terms of the GNU Affero General Public License as
 *     published by the Free Software Foundation, either version 3 of the
 *     License, or (at your option) any later version.
 *
 *     This program is distributed in the hope that it will be useful,
 *     but WITHOUT ANY WARRANTY; without even the implied warranty of
 *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *     GNU Affero General Public License for more details.
 *
 *     You should have received a copy of the GNU Affero General Public License
 *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
import { WithActions } from './../../../../mypay4-fe-common/src/lib/table/with-actions';

export class CapitoloRT extends WithActions {

  codUfficio: string;
  deUfficio: string;
  codCapitolo: string;
  deCapitolo: string;
  deAnnoEsercizio: string;
  codAccertamento: string;
  deAccertamento: string;
  numImporto: number;
}